requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
autopep8==1.7.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
Deprecated==1.2.13
Django==3.2
django-countries==7.3.2
django-debug-toolbar==3.4.0
django-filter==22.1
django-taggit==3.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Faker==14.2.0
idna==3.3
kombu==5.2.4
Markdown==3.4.1
packaging==21.3
Pillow==9.2.0
prompt-toolkit==3.0.31
pycodestyle==2.9.1
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytube==12.1.0
pytz==2022.2.1
redis==4.3.4
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.3.0
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
wget==3.2
wrapt==1.14.1
youtube-dl==2021.12.17
